Refactored data_worker.tsx
